+++ package.json
@@ -1,16 +1,20 @@
 {
   "name": "dc24bot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "bot.js",
+  "version": "2.0.0",
+  "description": "A Discord-Bot, made for Disconnect24.",
+  "main": "src/bot.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "TheMegaEagle",
+  "author": "TheMegaEagle and Apfel",
   "license": "ISC",
   "dependencies": {
     "discord.js": "^11.4.2",
+    "discord.js-commando": "^0.10.0",
+    "erlpack": "^0.1.2",
+    "mysql": "^2.16.0",
     "nodemon": "^1.18.9",
-    "request": "^2.88.0"
+    "request": "^2.88.0",
+    "uws": "^99.0.0"
   }
 }
