+++ src/commands/code/remove.js
