+++ src/commands/information/ping.js
@@ -1,0 +1,5 @@
+module.exports = {
+    run: function(bot, msg) {
+        msg.channel.send(`:ping_pong: Pong! **${Math.round(bot.ping)}ms**`)
+    }
+}
